+++ src/main/java/CPSF/com/demo/service/CamperPlaceService.java
@@ -6,6 +6,7 @@ import CPSF.com.demo.enums.Type;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public interface CamperPlaceService extends CRUDService<CamperPlace, CamperPlaceDTO>  {
@@ -16,7 +17,7 @@ public interface CamperPlaceService extends CRUDService<CamperPlace, CamperPlace
 
     boolean checkIsCamperPlaceOccupied(CamperPlace camperPlace, LocalDate checkin, LocalDate checkout, int reservationId);
 
-    void create(Type type, double price);
+    void create(Type type, BigDecimal price);
 
     void deleteByIndex(String index);
